requirements.txt
numpy
Pillow
tensorflow
keras
Keras-Preprocessing
fastapi
uvicorn
python-multipart